MIS-190 #comment Rename resolvers.ts to index.ts, making resolvers a module
